IT'S ALIVE
This is a small HTML base for presenting comps through a web browser.

The only adjustment that pages need is the inline CSS rule for the
image: margin-left. This must be half the width of the presented image.

This ensures that however your client is viewing the image, they will
see a nicely centered, pixel per pixel, comp.